Move notifier registration to chain worker
Finish preventing a race between adding a listener to the
`DeliveryNotifier` and notifying that messages have been delivered. Now
both steps are performed in the chain worker, which serializes them.
